+++ src/root/index.js
@@ -25,6 +25,7 @@ import Region from './views/region';
 import Country from './views/countries';
 import Deployments from './views/deployments';
 import Table from './views/table';
+import Covid19ThreeWSankey from './views/Covid19ThreeWSankey';
 
 import PerForms from './views/per-forms';
 import ViewPerForms from './views/view-per-forms';
@@ -89,6 +90,7 @@ const Root = () => (
     <Router>
       <Switch>
         <Route exact path='/' component={Home}/>
+        <Route exact path='/covid19-3w-sankey' component={Covid19ThreeWSankey}/>
         <Route exact path='/about' component={About}/>
         <PrivateRoute exact path='/account' component={Account}/>
         <PrivateRoute exact path='/account/password-change' component={PasswordChange}/>